Q4

+++ MAD 101 Q1 P4.cpp
@@ -16,10 +16,10 @@ int main() {
 	}
     int cnt = 0;
     for (int i = 0; i < primes.size(); i++){
-    	for (size_t j = i + 1; j < primes.size(); j++) {
+    	for (int j = i + 1; j < primes.size(); j++) {
             int pq = primes[i] * primes[j];
             if (pq > N) break;
-            for (size_t k = j + 1; k < primes.size(); k++) {
+            for (int k = j + 1; k < primes.size(); k++) {
                 int pqr = pq * primes[k];
                 if (pqr > N) break;
                 cnt++;
@@ -27,4 +27,5 @@ int main() {
         }
 	}
     cout << cnt;
+
 }
